requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp_socks==0.10.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
azure-ai-documentintelligence==1.0.2
azure-core==1.35.0
azure-identity==1.23.1
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
bottle==0.13.4
Brotli==1.1.0
browser-cookie3==0.20.1
cairocffi==1.7.1
CairoSVG==2.8.2
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
ChatterBot==1.2.3
click==8.2.1
cloudpathlib==0.21.1
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
cors==1.0.1
cryptography==45.0.5
cssselect2==0.8.0
curl_cffi==0.12.0
cymem==2.0.11
ddgs==9.4.3
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
frozenlist==1.7.0
future==1.0.0
g4f==0.5.7.6
gevent==25.5.1
googlemaps==4.10.0
gunicorn
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
jsonify==0.5
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.0
lz4==4.4.4
magika==0.6.2
mammoth==1.9.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.1.2
MarkupSafe==3.0.2
mathparse==0.1.2
mdurl==0.1.2
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
mss==10.0.0
multidict==6.6.3
murmurhash==1.0.13
nest-asyncio==1.6.0
nodriver==0.47.0
numpy==2.3.1
olefile==0.47
onnxruntime==1.22.1
openai==1.97.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pillow==11.3.0
platformdirs==4.3.8
plyer==2.1.0
preshed==3.0.10
primp==0.15.0
propcache==0.3.2
protobuf==6.31.1
proxy_tools==0.1.0
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
python-socks==2.7.1
pytz==2025.2
pywebview==5.4
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
srsly==2.5.1
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.47.2
sympy==1.14.0
thinc==8.3.6
tinycss2==1.4.0
tldextract==5.3.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
youtube-transcript-api==1.0.3
zope.event==5.1
zope.interface==7.2
flask-cors